+++ soxspipe/commonutils/__init__.py
@@ -8,3 +8,4 @@ from .create_dispersion_map import create_dispersion_map
 from .getpackagepath import getpackagepath
 from .detect_continuum import detect_continuum
 from . import polynomials
+from . import toolkit