+++ CHANGES.md
@@ -1,6 +1,9 @@
 # Release Notes
 
-y limits now added to the order table to show limits of order locations on detector
+* **FEATURE:** added a quicklook function in toolkit to quickly visualise a frame
+* **FEATURE:** added function in toolkit to unpack an order table into lists of coordinates, one list per order.
+* **FEATURE:** added a toolkit module for small functions used throughout soxspipe 
+* **ENHANCEMENT:** y limits now added to the order table to show limits of order locations on detector
 * **ENHANCEMENT:** added a fraction of a second tolerance when matching exptimes between darks and science/calibration frames 
 
 ## v0.4.1 - September 15, 2020